Suppression de dépendances non utilisées
